Load the Google Maps API globally
Since trying to use various loading components or hooks only leads to lots of problems. This is the simplest solution, which prevents the API from being loaded multiple times. Since checking that the API is loaded is tricky with Elmish, and since it normally loads very fast, we just assume that it is loaded when we need to draw a map.